requirements.txt
alembic==0.9.7
appnope==0.1.0
backcall==0.1.0
bcolz==1.2.1
bleach==3.0.2
boto==2.49.0
boto3==1.9.59
botocore==1.12.59
Bottleneck==1.2.1
bz2file==0.98
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
click==6.7
cloudpickle==0.8.0
contextlib2==0.5.5
cycler==0.10.0
cyordereddict==1.0.0
Cython==0.27.3
dask==1.1.3
decorator==4.3.0
defusedxml==0.5.0
Django==2.2
docutils==0.14
dyNET==2.1
empyrical==0.2.1
entrypoints==0.2.3
Flask==1.0.2
gensim==3.6.0
idna==2.7
image==1.5.27
intervaltree==2.1.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Logbook==0.12.5
lru-dict==1.1.6
lxml==4.3.2
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.0.2
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
multipledispatch==0.4.9
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.1
nltk==3.4
notebook==5.7.2
numexpr==2.6.4
numpy==1.16.2
olefile==0.46
pandas==0.19.2
pandas-datareader==0.6.0
pandocfilters==1.4.2
parso==0.3.1
patsy==0.5.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.4.2
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.0
pyparsing==2.3.0
python-dateutil==2.7.3
python-editor==1.0.4
pytz==2016.4
PyWavelets==1.0.2
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.20.1
requests-file==1.4.3
requests-ftp==0.3.1
s3transfer==0.1.13
scikit-image==0.14.2
scikit-learn==0.20.1
scipy==1.0.0
Send2Trash==1.5.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
smart-open==1.7.1
sortedcontainers==1.5.9
SQLAlchemy==1.2.2
sqlparse==0.3.0
statsmodels==0.8.0
tables==3.4.2
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
torch==1.0.1.post2
torchvision==0.2.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
wrapt==1.11.1
